Move out the auth package from internal
